Fixed the return type of a function, added comments to explain my code, and am resubmitting.

+++ ThreeMethodsV2/ThreeMethodsV2/Mathetizer.cs
@@ -11,7 +11,7 @@ public class Mathetizer
 
         public decimal Doublify(decimal a)
         {
-            return a * 2;
+            return (int)(a * 2);
         }
 
         public int Doublify(string a)
